Use navigator.share instead of copy if supported
If supported by the browser, trigger navigator.share on click of the
share button instead of copying the link and change the messages/hints
accordingly. This fixes the previous wrong use of navigator.share which
used an invalid argument format and was invoked without used interaction
which gets blocked by the browser.