Add build task to minify javascript
The code can be build via `yarn build` which minifies the code with Google's
closure compiler. We are compiling with the compiler's `SIMPLE` compilation
level as the `ADVANCED` level does not result in significantly smaller files
for this lib, especially when comparing gzipped size. One therefore does not
need to mind the special precautions that are required for the `ADVANCED`
level.

Resolves #4